+++ R/matrix.tools.R
@@ -95,18 +95,62 @@ n.observed_edges<- function(scan_list,diag=0){
 }
 
 #' Make Adjacency fit the selected mode
-#' From a directed adjacency matrix, make it fit the selected mode
+#' From a directed adjacency matrix, make it fit the selected mode.
 #'
 #' @param Adj an adjacency matrix
 #' @param mode Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.
 #'
-#' @return
+#' @return an adjacency matrix fitting the selected mode
 #' @export
 #'
 #' @examples
+#' set.seed(42)
+#'
+#' n<- 5;nodes<- letters[1:n];
+#' Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
+#' Adj[non.diagonal(Adj)]<- sample(0:30,n*(n-1),replace = TRUE)
+#' Adj<- iterate_scans(Adj,42,method="group",mode="directed",output = "adjacency",n.cores = 1)
+#' adjacency_mode(Adj,"max")
 adjacency_mode<- function(Adj,mode = c("directed", "undirected", "max","min", "upper", "lower", "plus")){
-  modew<- match.arg(mode)
+  mode<- match.arg(mode)
+  switch(mode,
+         "undirected" = ,
+         "max" = {
+           Adj<- ifelse(Adj>t(Adj),Adj,0)
+           Adj+t(Adj)
+         },
+         "min" = {
+           Adj<- ifelse(Adj<t(Adj),Adj,0)
+           Adj+t(Adj)
+         },
+         "plus" = Adj+t(Adj),
+         "directed" = ,
+         "upper" = ,
+         "lower" =  Adj
+  )
+}
 
+#' Make Binary Adjacency fit the selected mode
+#' From a directed binary adjacency matrix, make it fit the selected mode.
+#'
+#' @param Adj a binary adjacency matrix
+#' @param mode Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.
+#'
+#' @return a binary adjacency matrix fitting the selected mode
+#' @export
+#'
+#' @examples
+#' set.seed(42)
+#'
+#' n<- 5;nodes<- letters[1:n];
+#' Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
+#' Adj[non.diagonal(Adj)]<- sample(0:30,n*(n-1),replace = TRUE)
+#' Adj
+#'
+#' Adj<- do.scan(Adj,42)
+#' binary_adjacency_mode(Adj,"plus")
+binary_adjacency_mode<- function(Adj,mode = c("directed", "undirected", "max","min", "upper", "lower", "plus")){
+  mode<- match.arg(mode)
   switch(mode,
          "undirected" = ,
          "max" = ifelse(Adj+t(Adj)>=1,1,0), #conserve a connection between nodes if there's one in either directions (either adjacency triangle)
@@ -117,3 +161,4 @@ adjacency_mode<- function(Adj,mode = c("directed", "undirected", "max","min", "u
          "lower" =  Adj
   )
 }
+